drm: Uncomment checks on `oops_in_progress`
[Why]
Usually, these correspond to functions which must be aborted when there
is a panic in progress.

`oops_in_progress` is supported by linuxkpi, thus we can rely on it.

This should fix nested panics I saw because the some code verifies that
some locks were acquires but they were not (because they were correctly
skipped inside a panic). These lock assertions would cause another panic
on top of the one being processed.

[How]
We simply remove all the `#ifdef __FreeBSD__`.